+++ app/api/retrieval/process/route.ts
@@ -58,7 +58,6 @@ export async function POST(req: Request) {
       status: 200
     })
   } catch (error: any) {
-    console.log(error)
     const errorMessage = error.error?.message || "An unexpected error occurred"
     const errorCode = error.status || 500
     return new Response(JSON.stringify({ message: errorMessage }), {