+++ CHANGELOG.md
@@ -8,4 +8,4 @@
 
 * if tool returns an array with first element having a file_id attribute, then the return is considered a list of chunks
 
-* collections, assistants, collections and files can now be made public. Only the owner can edit
+* assistants, collections and files can now be made public. Only their owner can edit them